New DisplayRequest sample, moved Power and Display samples to common folder
