+++ src/SamplesApp/UITests.Shared/Windows_System/Display/DisplayRequest.xaml
@@ -1,0 +1,16 @@
+﻿<UserControl
+    x:Class="UITests.Shared.Windows_System.Display.DisplayRequest"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    d:DesignHeight="300"
+    d:DesignWidth="400">
+
+    <StackPanel>
+		<Button Click="RequestActive_Click">RequestActive()</Button>
+		<Button Click="RequestRelease_Click">RequestRelease()</Button>
+		<TextBlock x:Name="ActiveRequestCounter"/>
+	</StackPanel>
+</UserControl>
