+++ src/SamplesApp/UITests.Shared/Windows_System/Power/PowerManager.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="UITests.Shared.Windows_System_Power.PowerManager"
+    x:Class="UITests.Shared.Windows_System.Power.PowerManager"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
